requirements.txt
transformers>=4.21.0
torch>=2.0.0
numpy>=1.21.0
peft>=0.5.0
datasets>=2.0.0
accelerate>=0.20.0
evaluate>=0.4.0
seqeval>=1.2.2
pyyaml>=6.0.0
huggingface-hub>=0.16.0